requirements.txt
cosine_annealing_warmup==2.0
ftfy==6.1.1
lmdb==1.3.0
matplotlib==3.5.2
numpy==1.22.3
nystrom_attention==0.0.11
opencv_python==4.5.5.62
pandas==1.4.2
Pillow==9.1.0
pytorch_lightning==1.6.0
PyYAML==6.0
regex==2022.4.24
scikit_learn==1.0.2
seaborn==0.11.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.18.0
